:recycle: refactor: user enums 위치변경
