+++ src/main/java/umc/product/domain/member/entity/enums/LoginType.java
@@ -1,4 +1,4 @@
-package umc.product.domain.member.entity;
+package umc.product.domain.member.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
